randomQuestion problem
I suspect that there is a problem with the randomQuestion variable, when the variable is defined, say it rolled a 5, I would be stuck with [0][5],[1][5],[2][5] and so on and so forth. Going to create Random-Millionare branch to debug the issue. Will then push finished product to coding millionare to then continue forth with the project.